v1.5.25 TODO #27 Batch 2: HFSCX-256 in herradura.h + C known-answer test [19]
- herradura.h: add _HFSCX256_IV[32] constant and hfscx_256() (Merkle-Damgård
  hash on NL-FSCX v1, ISO 7816-4 padding + MD-strengthening length block)
- Herradura_tests.c [19]: four KAVs (empty, 0x61, 33*A, collision-resistance)
  cross-checked against Python hfscx_256(); benchmarks renumbered [19]-[28] →
  [20]-[29]; version banner bumped to v1.5.25
- TODO.md: Batch 2 marked complete
